engines/ciem/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
psycopg2-binary>=2.9.9
pydantic>=2.5.0
kubernetes>=28.1.0
pyyaml>=6.0
# AWS
boto3>=1.34.0
# Azure (optional — installed when Azure accounts onboarded)
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
# GCP (optional — installed when GCP accounts onboarded)
google-cloud-storage>=2.14.0
google-auth>=2.25.0
# OCI (optional)
oci>=2.120.0
# IBM (S3-compatible)
ibm-cos-sdk>=2.13.0
